+++ loggingFunctions/logBulkDeletedMessage.js
