+++ loggingFunctions/logDeletedMessage.js
