+++ loggingFunctions/logEditedMessage.js
