+++ loggingFunctions/logSentMessage.js
