+++ loggingFunctions/logVcConnection.js
